17_12_3_32

+++ matchmaker/src/main/java/ru/atom/dao/GameDao.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 import ru.atom.model.Game;
 
-public interface GameDao extends CrudRepository<Game, Integer> {
+public interface GameDao extends CrudRepository<Game, Long> {
 }
